+++ core/src/main/java/feign/Contract.java
@@ -202,19 +202,6 @@ public interface Contract {
     protected abstract boolean processAnnotationsOnParameter(
         MethodMetadata data, Annotation[] annotations, int paramIndex);
 
-    /**
-     * @deprecated dead-code will remove in feign 10
-     */
-    @Deprecated
-    // deprecated as only used in a sub-type
-    protected Collection<String> addTemplatedParam(Collection<String> possiblyNull, String name) {
-      if (possiblyNull == null) {
-        possiblyNull = new ArrayList<String>();
-      }
-      possiblyNull.add(String.format("{%s}", name));
-      return possiblyNull;
-    }
-
     /** links a parameter name to its index in the method signature. */
     protected void nameParam(MethodMetadata data, String name, int i) {
       Collection<String> names =